clang-tidy: (WIP) bugprone-too-small-loop-variable findings 2
Reviewed-on: https://gerrit.libreoffice.org/63241
Tested-by: Jenkins

+++ sal/qa/rtl/strings/test_oustring_convert.cxx
@@ -170,7 +170,7 @@ void test::oustring::Convert::convertToString()
               OUSTRING_TO_OSTRING_CVTFLAGS,
               "A?B",
               "A?B" } };
-    for (unsigned int i = 0; i < SAL_N_ELEMENTS(aTests); ++i)
+    for (size_t i = 0; i < SAL_N_ELEMENTS(aTests); ++i)
         testConvertToString(aTests[i]);
 }
